split into commands
