+++ wandb/cli/commands/launch/__init__.py
@@ -1,0 +1,12 @@
+from wandb.cli.commands.launch.agent import launch_agent
+from wandb.cli.commands.launch.job import job
+from wandb.cli.commands.launch.launch import launch
+from wandb.cli.commands.launch.sweep import launch_sweep, scheduler
+
+__all__ = (
+    "job",
+    "launch_sweep",
+    "scheduler",
+    "launch_agent",
+    "launch",
+)
